Finishing touches

+++ mono_detector/include/mono_detector/types.hpp
@@ -40,15 +40,15 @@ struct CameraModel {
   void fromCameraInfo(sensor_msgs::msg::CameraInfo camera_info) {
     if (camera_info.distortion_model == sensor_msgs::distortion_models::PLUMB_BOB) {
       distortion_model = DistortionModel::PINHOLE;
-      distortion_parameters = cv::Mat(1, 5, CV_32F, camera_info.d.data()).clone();
+      distortion_parameters = cv::Mat(1, 5, CV_64F, camera_info.d.data()).clone();
     } else if (camera_info.distortion_model == sensor_msgs::distortion_models::EQUIDISTANT) {
       distortion_model = DistortionModel::FISHEYE;
-      distortion_parameters = cv::Mat(1, 4, CV_32F, camera_info.d.data()).clone();
+      distortion_parameters = cv::Mat(1, 4, CV_64F, camera_info.d.data()).clone();
     } else {
       throw std::runtime_error("Unsupported distortion model: " + camera_info.distortion_model);
     }
 
-    camera_matrix = cv::Mat(3, 3, CV_32F, camera_info.k.data()).clone();
+    camera_matrix = cv::Mat(3, 3, CV_64F, camera_info.k.data()).clone();
   }
 };
